Some User controllers have been updated
The follow user controller has been updated to check whether the user to follow has a private or public profile and its respective functionality. The controller for changing the request status has been updated to optimize how the request status and the following and notification lists for the sending user are updated. Finally, a controller to cancel the follow request and remove the notification from the notification list of the user to whom the request is being canceled has been implemented